moved and renamed some middlewares
- AuthMiddleware
- InitContextMiddleware
- TokenAuthMiddleware
