+++ internal/api/v1/middleware/token.go
@@ -1,0 +1,80 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package middleware
+
+import (
+	"net/http"
+
+	"github.com/epinio/epinio/helpers/authtoken"
+	"github.com/epinio/epinio/internal/api/v1/response"
+	"github.com/epinio/epinio/internal/auth"
+	"github.com/epinio/epinio/internal/cli/server/requestctx"
+	apierrors "github.com/epinio/epinio/pkg/api/core/v1/errors"
+	"github.com/gin-gonic/gin"
+	"github.com/golang-jwt/jwt/v4"
+)
+
+// TokenAuth middleware is used to authenticate a user from a 'authtoken'
+// It's used when trying to establish a websocket connections for authenticated users.
+func TokenAuth(ctx *gin.Context) {
+	logger := requestctx.Logger(ctx.Request.Context()).WithName("TokenAuth")
+	logger.V(1).Info("Authtoken authentication")
+
+	token := ctx.Query("authtoken")
+	claims, err := authtoken.Validate(token)
+	if err != nil {
+		apiErr := apierrors.NewAPIError("unknown token validation error", http.StatusUnauthorized)
+		if ve, ok := err.(*jwt.ValidationError); ok {
+			if ve.Errors&jwt.ValidationErrorMalformed != 0 {
+				apiErr.Title = "malformed token format"
+
+			} else if ve.Errors&(jwt.ValidationErrorExpired|jwt.ValidationErrorNotValidYet) != 0 {
+				apiErr.Title = "token expired"
+
+			} else {
+				apiErr.Title = "cannot handle token"
+			}
+		}
+
+		// detailed log message
+		logger.V(2).Info(apiErr.Title, "error", err.Error())
+		// not too specific log message for unauthorized client
+		response.Error(ctx, apiErr)
+		ctx.Abort()
+		return
+	}
+
+	authService, err := auth.NewAuthServiceFromContext(ctx)
+	if err != nil {
+		response.Error(ctx, apierrors.InternalError(err))
+		ctx.Abort()
+		return
+	}
+
+	// find the user and add it in the context
+	users, err := authService.GetUsers(ctx)
+	if err != nil {
+		response.Error(ctx, apierrors.InternalError(err))
+		ctx.Abort()
+		return
+	}
+
+	for _, user := range users {
+		if user.Username == claims.Username {
+			newCtx := ctx.Request.Context()
+			newCtx = requestctx.WithUser(newCtx, user)
+			ctx.Request = ctx.Request.Clone(newCtx)
+
+			break
+		}
+	}
+}
